+++ server/views/pages/bookers/booker/bookerDetails.njk
@@ -1,0 +1,59 @@
+{% extends "partials/layout.njk" %}
+{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
+{% from "govuk/components/button/macro.njk" import govukButton %}
+
+{% set pageHeaderTitle = "Manage bookers" %}
+{% set pageTitle = applicationName + " - " + pageHeaderTitle %}
+{% set activePrimaryNav = "bookers" %}
+
+{% block content %}
+{{ super() }}
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+
+    <h1 class="govuk-heading-l">{{ pageHeaderTitle }}</h1>
+
+    <h2 class="govuk-heading-m">Booker details</h2>
+
+    {{ govukSummaryList({
+      rows: [
+        {
+          key: {
+            text: "Booker"
+          },
+          value: {
+            text: booker.email
+          }
+        },
+        {
+          key: {
+            text: "Prisoner"
+          },
+          value: {
+            text: booker.permittedPrisoners[0].prisonerId
+          } if booker.permittedPrisoners | length else {
+            html: '<a href="/bookers/booker/prisoner/add">Add a prisoner</a>'
+          }
+        }
+      ]
+    }) }}
+
+    {% if booker.permittedPrisoners | length %}
+      <h2 class="govuk-heading-m">Clear booker details</h2>
+      <p>This will remove prisoner and visitor records for this booker.</p>
+      <p>The email, booker reference and GOV.UK One Login identifier will be kept.</p>
+
+      <form action="/bookers/booker/clear-details" method="POST" novalidate>
+        <input type="hidden" name="_csrf" value="{{ csrfToken }}">
+        {{ govukButton({
+          text: "Clear booker details",
+          classes: "govuk-button--warning",
+          preventDoubleClick: true
+        }) }}
+      </form>
+    {% endif %}
+  </div>
+</div>
+
+{% endblock %}
