--- server/views/pages/bookers/bookerDetails.njk
@@ -1,22 +1,0 @@
-{% extends "partials/layout.njk" %}
-
-{% set pageHeaderTitle = "Manage bookers" %}
-{% set pageTitle = applicationName + " - " + pageHeaderTitle %}
-{% set activePrimaryNav = "bookers" %}
-
-{% block content %}
-{{ super() }}
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-full">
-
-    <h1 class="govuk-heading-l">{{ pageHeaderTitle }}</h1>
-
-    <h2 class="govuk-heading-m">Booker details</h2>
-
-    <pre>{{ booker | dump(2) }}</pre>
-
-  </div>
-</div>
-
-{% endblock %}
